+++ Changelog.md
@@ -14,6 +14,17 @@
 ## Documentation
 -->
 
+# 2.3.3
+
+## Steps
+
+* `OpenROAD.Floorplan`
+
+  * Fixed an issue in `FP_SIZING`: `absolute` mode where if the die area's
+    x0 > x1 or y0 > y1, the computed core area would no longer fit in the die
+    area. Not that we recommend you ever do that, but technically OpenROAD
+    allows it.
+
 # 2.3.2
 
 ## Steps